requirements.txt
PyYAML==6.0.1
imutils==0.5.4
numpy==1.26.0
opencv-python==4.8.0.76